Added Build system for static site generator
Added SSG for the main index page, the main blog page, and each of the
blog subpages. The main index and blog page were created with
handlebars, using a Rust binary to coordinate information, and a TOML
file where the information. The blog subpages are written in markdown
for easy editing, and then are appended to a HTML template for format.

It was difficult for me to justify updating my website since the
html is sometimes hard to modify properly, so I decided to created a
static site generator to simplify my life.

Instructions:

- To modify main page information, modify the TOML file

- Main blog page information links are automatically created based on
  available blogs

- Blogs are written in Markdown, and edited using retext for the preview
  feature