Add link to substack newsletter on all navbars
I modified all the handlebars templates and the bot.txt append template
to use the substack link instead of the relative link to
newsletter.html. In theory the newsletter html page is still available,
but should not be accessed.

I did this because it's more simple then rewritting all my newsletter in
a website format. The substack format is really nice, and essentially
this just redirects people reading to my substack page.