Fix Resume link, and other style changes
Changed the old useless newsletter page into a new resume page. Changed
the style for a tags to be a little lighter, and also did some other
style changes in both style.css files for the index and blog file.

I did this because I want people on my website to have a better
expirience, and not to be suprised that the website is "incomplete" and
not working. Now every feature is good, and when I eventually make my
resume I'll add it to the website.