Remove quiet compilation of both rust binary
Removed the -q flag on both cargo builds.

I did this because running the `make index` and `make blog` commands
apeared to be failing from my new macbook, when in fact they only
apeared to be failing because the compilation for the two rust binary
were taking longer then anticipated. This way it is made clear for the
person running the code that the script is working.