Revamp blog build system to add title
I now use a rust binary with handlebars for compiling each blog. The
system is kinda wacky but it works. Done > Perfect.

Did this because the SEO for my blogs was trash. Adding the meta tags
for desc and a good title tag will go a long way toward increasing my
  SEO.